refactor of resources hooks
