--- modules/040-control-plane-manager/hooks/resources_budget.go
@@ -1,137 +1,0 @@
-/*
-Copyright 2026 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package hooks
-
-import (
-	"fmt"
-
-	"github.com/flant/addon-operator/pkg/module_manager/go_hook"
-	"github.com/flant/addon-operator/sdk"
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-)
-
-const (
-	controlPlanePercent     = 40                // %
-	configEveryNodeMilliCPU = 300               // 0.3 Cpu
-	configEveryNodeMemory   = 512 * 1024 * 1024 // 512Mb
-	hardLimitMilliCPU       = 4 * 1000          // 4 Cpu
-	hardLimitMemory         = 8 * 1024 * 1024 * 1024
-
-	// Asymmetric deadband for resource request updates (autotune + legacy
-	// combined-budget calculate). Go constants, not config-values.
-	// Change is significant only when delta > raiseThreshold or
-	// delta < -lowerThreshold (delta = (rec - applied) / applied).
-	raiseThreshold = 0.20 // +20%
-	lowerThreshold = 0.30 // −30%
-
-	// Minimum kubelet reservation we account for, regardless of what the kubelet
-	// has actually reported on Node.Status.Allocatable at the moment the hook
-	// runs. The hook uses Capacity (immutable) and subtracts max(actual kubelet
-	// reservation, this floor) so the result is identical before and after the
-	// kubelet finishes initialising — which avoids a second hook run later that
-	// would re-render every control-plane static-pod manifest and cascade-restart
-	// kube-apiserver/etcd/kcm/ks right in the middle of Deckhouse install.
-	kubeletResourceReservationMemoryFloor = 900 * 1024 * 1024 // 900 MiB
-	kubeletResourceReservationCPUFloor    = 100               // 0.1 cpu
-)
-
-type Node struct {
-	CapacityMilliCPU    int64
-	CapacityMemory      int64
-	AllocatableMilliCPU int64
-	AllocatableMemory   int64
-}
-
-func applyNodesResourcesFilter(obj *unstructured.Unstructured) (go_hook.FilterResult, error) {
-	node := &v1.Node{}
-	err := sdk.FromUnstructured(obj, node)
-	if err != nil {
-		return nil, fmt.Errorf("from unstructured: %w", err)
-	}
-
-	n := &Node{
-		AllocatableMilliCPU: node.Status.Allocatable.Cpu().MilliValue(),
-		AllocatableMemory:   node.Status.Allocatable.Memory().Value(),
-		CapacityMilliCPU:    node.Status.Capacity.Cpu().MilliValue(),
-		CapacityMemory:      node.Status.Capacity.Memory().Value(),
-	}
-	// Test fixtures and very early node objects may not report Capacity yet —
-	// fall back to Allocatable. The downstream logic treats `Capacity == Allocatable`
-	// as `kubelet has not subtracted its reservation yet` and applies the floor.
-	if n.CapacityMilliCPU == 0 {
-		n.CapacityMilliCPU = n.AllocatableMilliCPU
-	}
-	if n.CapacityMemory == 0 {
-		n.CapacityMemory = n.AllocatableMemory
-	}
-
-	return n, nil
-}
-
-// effectiveMasterResources returns the per-node usable CPU/memory budget the
-// control-plane allocation can be carved out of. Computed from Node.Status.Capacity
-// (immutable for the lifetime of the node) minus max(actual kubelet reservation,
-// our floor). The result is stable across the kubelet warm-up window, so the
-// hook output does not flip a few minutes into the bootstrap.
-func effectiveMasterResources(n *Node) (int64, int64) {
-	cpuReservation := n.CapacityMilliCPU - n.AllocatableMilliCPU
-	if cpuReservation < kubeletResourceReservationCPUFloor {
-		cpuReservation = kubeletResourceReservationCPUFloor
-	}
-	memReservation := n.CapacityMemory - n.AllocatableMemory
-	if memReservation < kubeletResourceReservationMemoryFloor {
-		memReservation = kubeletResourceReservationMemoryFloor
-	}
-	return n.CapacityMilliCPU - cpuReservation, n.CapacityMemory - memReservation
-}
-
-// minMasterNodeBudget returns the control-plane allocatable budget of the
-// weakest master: effectiveMasterResources(minNode) − configEveryNode.
-func minMasterNodeBudget(nodes []Node) (int64, int64, bool) {
-	if len(nodes) == 0 {
-		return 0, 0, false
-	}
-
-	discoveryMasterNodeMilliCPU := int64(hardLimitMilliCPU)
-	discoveryMasterNodeMemory := int64(hardLimitMemory)
-
-	for i := range nodes {
-		effCPU, effMem := effectiveMasterResources(&nodes[i])
-		if effCPU < discoveryMasterNodeMilliCPU {
-			discoveryMasterNodeMilliCPU = effCPU
-		}
-		if effMem < discoveryMasterNodeMemory {
-			discoveryMasterNodeMemory = effMem
-		}
-	}
-
-	return discoveryMasterNodeMilliCPU - configEveryNodeMilliCPU,
-		discoveryMasterNodeMemory - configEveryNodeMemory,
-		true
-}
-
-// significantResourceChange reports whether rec differs from applied enough to
-// commit an update under the asymmetric deadband. applied <= 0 always accepts
-// a positive recommendation (first commit).
-func significantResourceChange(rec, applied int64) bool {
-	if applied <= 0 {
-		return rec > 0
-	}
-	delta := float64(rec-applied) / float64(applied)
-	return delta > raiseThreshold || delta < -lowerThreshold
-}
